refactor: 전략 패키지 파일 분리
- 전략 패키지 관련 파일을 /domain/strategy로 분리
