+++ src/main/java/domain/strategy/PurchaseStrategy.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.strategy;
 
 import java.util.List;
