+++ src/main/java/domain/strategy/RandomPurchaseStrategy.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.strategy;
 
 import java.util.ArrayList;
 import java.util.Collections;
